Update index.html error Fixing
Error: Element div not allowed as child of element ul in this context. (Suppressing further errors from this subtree.)

From line 44, column 19; to line 44, column 44

          <div class="search-right">↩     

Contexts in which element div may be used:
Where flow content is expected.
As a child of a dl element.
Content model for element ul:
Zero or more li and script-supporting elements.